chore: realocate sign in and sign up pages
